Convert table model to outer class
add class comment
